basic implementation for SeRo Systems API (#109)
An example program to display receivers information
Auto-download the ca.crt file to the config directory
Define filters in the configuration file
Refactoring to distinguish Source and Sensor
Refactor the use of metadata (requires a HashMap rather than an index)

+++ crates/decode1090/src/main.rs
@@ -45,7 +45,7 @@ struct Options {
 #[derive(Serialize, Deserialize)]
 struct JSONEntry {
     timestamp: f64,
-    rssi: Option<f64>, // from older format
+    rssi: Option<f32>, // from older format
     #[serde(
         serialize_with = "rs1090::decode::as_hex",
         deserialize_with = "rs1090::decode::from_hex"
